requirements.txt
memory_profiler
requests==2.25.1
psycopg2-binary==2.8.6